# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saknussemm",
    "description": "Post-OCR text correction of ALTO/PAGE XML files using LLM providers.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/jobs": {
      "post": {
        "tags": ["jobs"],
        "summary": "Create Job",
        "description": "Upload ALTO/PAGE XML files and start a correction job.\n\nUpload concurrency is reserved by ``UploadAdmissionMiddleware``\nBEFORE the multipart body is read: by the time this handler runs,\nFastAPI has already parsed (and spooled) the files, so a slot check\nhere would fire after the full upload cost was paid. The whole\nhandler executes inside the middleware's reservation.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_api_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get Job",
        "description": "Poll the status of a correction job.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_job_api_jobs_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "geometric_pairing": {
            "type": "boolean",
            "title": "Geometric Pairing",
            "default": true
          }
        },
        "type": "object",
        "required": ["files", "provider", "api_key", "model"],
        "title": "Body_create_job_api_jobs_post"
      },
      "CreateJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Token"
          },
          "events_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Url"
          }
        },
        "type": "object",
        "required": ["job_id"],
        "title": "CreateJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "started",
          "running",
          "completed",
          "completed_with_fallbacks",
          "failed",
          "cancel_requested",
          "cancelled"
        ],
        "title": "JobStatus",
        "description": "Lifecycle state of a correction job, surfaced to API clients."
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "total_lines": {
            "type": "integer",
            "title": "Total Lines",
            "default": 0
          },
          "lines_modified": {
            "type": "integer",
            "title": "Lines Modified",
            "default": 0
          },
          "chunks_total": {
            "type": "integer",
            "title": "Chunks Total",
            "default": 0
          },
          "retries": {
            "type": "integer",
            "title": "Retries",
            "default": 0
          },
          "fallbacks": {
            "type": "integer",
            "title": "Fallbacks",
            "default": 0
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": ["job_id", "status"],
        "title": "JobStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}